- Added new French localization. Thanks Ronald Leroux! - AppTrap background process now, by default, compiles into 32/64 bit Universal (i386, x86_64 and ppc). - Modified UKKQueue to compile without warnings. I simply changed one of the values from being an int to an NSInteger. Same thing with ATApplicationController.m in the moveCurrentItemsToTrash: method. - Added Ronald Leroux as a localizer to Read Me.rtf

+++ AppTrap/UKKQueue/UKKQueue.m
@@ -269,7 +269,8 @@ static UKKQueue * gUKKQueueSharedQueueSingleton = nil;
 
 -(void) removePathFromQueue: (NSString*)path
 {
-    int		index = 0;
+    //int		index = 0;
+	NSInteger index = 0;
     int		fd = -1;
     
     AT_SYNCHRONIZED( self )
